Revert "Merge pull request #198 from hashicorp/chore/terraform-plugin-framework@v0.4.2"
This reverts commit 237544d42b8f21ab9db0e96cfef0126d565b59ac, reversing
changes made to 041f107c83e8eae76d0c1ba3b3a425a2d021044b.

+++ internal/validate/for_each.go
@@ -40,7 +40,7 @@ func (validator arrayForEachValidator) Validate(ctx context.Context, request tfs
 	}
 
 	for i, e := range l.Elems {
-		elemPath := request.AttributePath.WithElementKeyInt(i)
+		elemPath := request.AttributePath.WithElementKeyInt(int64(i))
 
 		elemRequest := tfsdk.ValidateAttributeRequest{
 			AttributePath:   elemPath,
